add -cb option
Now we can specify a callback proc that will be called on every conf
key value set/append.
Option accept a list, where 1 element is proc name and 2 element is
priv data(which can be omitted). If -cb option is specified, then
all load proc return a list where 1 element is a conf dict and 2
element is a priv data.
A callback accepts the next arguments in order:
- ctx var name(dict where a priv data is saved in "priv" key)
- conf var name(dict)
- operation name (=S, =L, +=S, +=L, ?=S, ?=L - where S is string and L
  is list)
- key names var name(list with full key path)
- value var name(list)